fix: task and docker
