+++ build_version
@@ -1,2 +1,2 @@
-APP_VERSION=0.9.0
+APP_VERSION=0.9.1
 APP_NAME=sales_bot
