+++ build_version.bak
@@ -1,2 +1,2 @@
-APP_VERSION=0.9.0-29655
+APP_VERSION=0.9.1
 APP_NAME=sales_bot
