+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "sales_bot"
-version = "0.9.0"
+version = "0.9.1"
 description = ""
 authors = ["Anton Zimin <zimkaa87@gmail.com>"]
 readme = "README.md"
