+++ src/config/project_info.py
@@ -1,5 +1,3 @@
-import sys
-
 import tomllib
 
 
@@ -9,9 +7,7 @@ def get_name() -> str:
     return data["tool"]["poetry"]["name"]
 
 
-def main() -> None:
-    sys.stdout.write(f"{get_name()}\n")
-
-
-if __name__ == "__main__":
-    main()
+def get_version() -> str:
+    with open("pyproject.toml", "rb") as f:  # noqa: PTH123
+        data = tomllib.load(f)
+    return data["tool"]["poetry"]["version"]
